refactor: line_line_intersection_raw を common モジュールに集約
レビュー指摘対応 (comment 3177672290 / 3177672294 / 3177672299 / 3177672307):
- intersection/common.rs に line_line_intersection_raw を新規配置
- 分母ゼロ/極小チェック追加（分母 <= tolerance^2 の場合 None）
- pair_base と primitive_3d から super::super::common に参照（依存逆転解消）
- primitive_3d 側は pub(in crate::intersection) に可視性制限
- shared.rs から line_line_intersection_raw 削除
